some 64-bit warnings (and bugs, really) are fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@16719 9d14d920-75d5-0310-8bd7-a52572981aee

+++ suites/rcftest/rcftest.c
@@ -41,7 +41,7 @@ int
 main()
 {
     char ta[32];
-    int  len = sizeof(ta);
+    size_t len = sizeof(ta);
     int  sid;
     
     printf("Starting test\n");
