+++ suites/selftad/ipstack/tcp_simple.c
@@ -65,7 +65,7 @@ main(int argc, char *argv[])
     char ta[32];
     char *agt_a = ta;
     char *agt_b;
-    int  len = sizeof(ta);
+    size_t  len = sizeof(ta);
 
     char path[1000];
